+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 - Deprecate `configunmarshaler` package, move it to internal (#5151)
 - Deprecate all API in `model/semconv`. The package is moved to a new `semcomv` module (#5196)
+- Deprecate `p<signal>otlp.Request.Set<Logs|Metrics|Traces>` (#5234)
+  - `plogotlp.Request.SetLogs` func is deprecated in favor of `plogotlp.NewRequestFromLogs`
+  - `pmetricotlp.Request.SetMetrics` func is deprecated in favor of `pmetricotlp.NewRequestFromMetrics`
+  - `ptraceotlp.Request.SetTraces` func is deprecated in favor of `ptraceotlp.NewRequestFromTraces`
 - `pmetric.NumberDataPoint.ValueType()` now returns new type `NumberDataPointValueType` (#5233)
   - `pmetric.MetricValueType` is deprecated in favor of `NumberDataPointValueType`
   - `pmetric.MetricValueTypeNone` is deprecated in favor of `NumberDataPointValueTypeNone`